Overnight high and low strategy mockup. kOvernightExtremeTest.cs
I would need to manually program the entry and exit for the time being because it is an indicator. But that will give me so exposure to create my own strategy reporting vs. Ninjatrader analyser.

Currently the strategy is not working, and I am using it as a tool to generate data for analysis.